Change search-facet-worktype component so it is building the types itself.
