+++ addon/components/search-facet-worktype/template.hbs
@@ -1,7 +1,7 @@
 {{!Copied from Ember-SHARE}}
 <div class='filter-block'>
     {{search-facet-worktype-hierarchy state=state defaultCollapsed=true
-        data=data onClick=(action 'toggle')}}
+        data=processedTypes onClick=(action 'toggle')}}
     <div class='indent row'>
         {{search-facet-worktype-button type='creative work' collapsible=false
                 selectedTypes=selected onClick=(action 'toggle') toggleState=toggleState
